deactivating guards added

+++ src/app/workout/workout-create/exercise-item/exercise-item.component.ts
@@ -1,5 +1,6 @@
-import { Component, OnInit, Input } from '@angular/core';
+import { Component, OnInit, Input, OnDestroy } from '@angular/core';
 import { Router } from '@angular/router';
+import { Subscription } from 'rxjs';
 
 import { ExerciseService } from '../../../shared/services/exercise.service';
 import { Exercise } from '../../../shared/models/exercise.model';
@@ -9,8 +10,10 @@ import { Exercise } from '../../../shared/models/exercise.model';
   templateUrl: './exercise-item.component.html',
   styleUrls: ['./exercise-item.component.scss']
 })
-export class ExerciseItemComponent implements OnInit {
+export class ExerciseItemComponent implements OnInit, OnDestroy {
   @Input() exercise: Exercise;
+  editMode: boolean;
+  editModeSubscription: Subscription;
 
   constructor(
     private exerciseService: ExerciseService,
@@ -18,11 +21,19 @@ export class ExerciseItemComponent implements OnInit {
   ) { }
 
   ngOnInit() {
+    this.editModeSubscription = this.exerciseService.exerciseEditModeChanged
+      .subscribe((editMode) => this.editMode = editMode);
   }
 
   onRemoveExercise() {
-    this.exerciseService.removeCurrentExercise(this.exercise.id);
-    this.router.navigate(['workout', 'exercises']);
+    if (!this.editMode) {
+      this.exerciseService.removeCurrentExercise(this.exercise.id);
+      this.router.navigate(['workout', 'exercises']);
+    }
+  }
+
+  ngOnDestroy() {
+    this.editModeSubscription.unsubscribe();
   }
 
 }
